// src/app/api/zona/models/zona-torrent.model.ts
export class ZonaTorrent {
  torrent_download_link: string;
  private: boolean;
  size_bytes: number;
  peers: number;
  language: string;
  translate_info: string;
  resolution: string;
  trailer: boolean;
  subtitles_parser: string;
  loading_time_sum: number;
  audio_quality_id: number;
  id: number;
  languages_parser: string;
  kinopoisk_id: number;
  seeds: number;
  loading_fail_count: number;
  quality_id: number;
  files: string;
  loading_success_count: number;
  hash: string;
  _health;
}